ref: graph restructurisation
