+++ src/main/kotlin/model/graph/UndirectedGraph.kt
@@ -1,11 +1,12 @@
-package model.graph.undirected
+package model.graph
 
+import graph.Graph
 import model.graph.edges.Edge
 
-class UndirectedUnweightedGraph<V> : UndirectedGraph<V>() {
+class UndirectedGraph<V> : Graph<V>() {
     override fun addEdge(from: V, to: V, weight: Int) {
-        val edge1 = Edge(from, to)
-        val edge2 = Edge(to, from)
+        val edge1 = Edge(from, to, weight)
+        val edge2 = Edge(to, from, weight)
         graph[from]?.add(edge1) ?: { graph[from] = mutableListOf(edge1) }
         graph[to]?.add(edge2) ?: { graph[to] = mutableListOf(edge2) }
     }
