--- src/main/kotlin/model/graph/directed/DirectedGraph.kt
@@ -1,5 +1,0 @@
-package model.graph.directed
-
-import graph.Graph
-
-abstract class DirectedGraph<V> : Graph<V>()
