--- src/main/kotlin/model/graph/undirected/UndirectedGraph.kt
@@ -1,6 +1,0 @@
-package model.graph.undirected
-
-import graph.Graph
-
-abstract class UndirectedGraph<V> : Graph<V>() {
-}
